requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
httpx
pytest
pytest-asyncio
langchain
langchain-anthropic
langchain-groq
pydantic
aiosqlite
python-multipart
